Move old database folder to old_database to avoid collisions
