+++ pykollib/request/HermitRequest.py
@@ -1,5 +1,5 @@
 import pykollib.Error as Error
-from pykollib.database import ItemDatabase
+from pykollib.old_database import ItemDatabase
 from pykollib.util import ParseResponseUtils
 from pykollib.pattern import PatternManager
 from pykollib.request.GenericRequest import GenericRequest
