+++ pykollib/request/MakePasteRequest.py
@@ -1,6 +1,6 @@
 import pykollib.Error as Error
 from .GenericRequest import GenericRequest
-from pykollib.database import ItemDatabase
+from pykollib.old_database import ItemDatabase
 from pykollib.pattern import PatternManager
 from pykollib.util import ParseResponseUtils
