+++ pykollib/request/MalusRequest.py
@@ -1,5 +1,5 @@
 from .GenericRequest import GenericRequest
-from pykollib.database import ItemDatabase
+from pykollib.old_database import ItemDatabase
 from pykollib.pattern import PatternManager
