+++ pykollib/request/WokRequest.py
@@ -1,5 +1,5 @@
 import pykollib.Error as Error
-from pykollib.database import ItemDatabase
+from pykollib.old_database import ItemDatabase
 from pykollib.pattern import PatternManager
 from pykollib.request.GenericRequest import GenericRequest
